[137] Add style to buttons
