+++ src/components/FreakDetails/FreakDetails.scss
@@ -1,7 +1,7 @@
 .freak-details {
   border-radius: 3.5rem;
   text-align: center;
-  margin: 10rem auto;
+  margin: 10rem auto 0;
   border: 2rem solid;
   background: #002430;
   color: white;
