+++ src/pages/ViewFreakPage/ViewFreakPage.scss
@@ -1,0 +1,7 @@
+.view-freak {
+  &__buttons {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+  }
+}
